library_api/serializers/rating_serializer.py
from rest_framework import serializers
from ..models import Rating

class RatingSerializer(serializers.ModelSerializer):
    user = serializers.ReadOnlyField(source='user.username')

    class Meta:
        model = Rating
        fields = '__all__'
