library_api/requirements.txt
Django
djangorestframework
drf-yasg
django-filter